Only perform asynchronous, save-time hashing
The original default method of hashing passwords during the `setter`
method was not ideal for node async i/o. The benefit of doing it at
set-time was to immediately hash the password, to prevent any leaking.
To add some of this security to the asynchronous method, we use the
bookshelf `virtuals` plugin, and never set `password` on the model
`attributes` hash.  This will prevent the password from being exposed
during a toJSON call (or other).